Switch formik map to leaflet
